+++ src/cqfantasy/house/__init__.py
@@ -1,0 +1,3 @@
+from .Body import Body
+from .Roof import Roof
+from .House import House